requirements.txt
decord==0.6.0
gradio==3.39.0
iopath==0.1.9
matplotlib==3.7.0
numpy==1.24.3
omegaconf==2.3.0
opencv_python==4.7.0.72
opencv_python_headless==4.5.5.64
pandas==2.0.1
Pillow==9.4.0
Pillow==10.0.0
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.11.1
skimage==0.0
tabulate==0.9.0
timm==0.4.12
torch==2.0.0+cu117
torchvision==0.15.1+cu117
tqdm==4.64.1
transformers==4.30.2
webdataset==0.2.48
